[ refactor ] Refactor AppendixSection
Changes:
- Change GlossarySection to AppendixSection.
- Update nav to have appendix instead of glossary.

Reference:
https://trello.com/c/XtTe3JPV/12-create-site-templates
